Extract location options from mapping code
